Changes in network state revalidate sources rhbz#1270354
When altering the network state, it may result in selected
network sources becoming inaccessable due to typos in DNS
or GATEWAY.

With this patch, after the network spoke completes it tells
the software payload to revalidate its state.  If the state
becomes invalid, the source spoke errors will guide towards
what is wrong.