Raise NonInteractive exception in Hubs event loop
Raise exception to stop installation when user interaction is required.

Before this commit we used to run event loop in every Hub (SummaryHub,
ProgressHub) but we never turn it off and all the event loops worked
with the same queue of events in parallel.
Now when new Hub is created it first disable old running event loop and
then it enables its own loop.