Support non interactive mode in standalone spokes
Code logic for reaction on stop of automatic kickstart installation was
moved to the new method autoinstall_stopped(reason) in
pyanaconda.ui.gui.helpers source file. Reason of the move is because we
need to call this logic from Hub and from GraphicalUserInterface too.
GraphicalUserInterface is a place where Standalone spokes are tested on
user interaction.