Disable test-install in Makefile
Purpose of this test is to find out if all dependencies can be resolved
on a target Fedora release. However, we don't want to block other tests
because there is a missing dependency on Fedora.

This should be run as separate CI job and it should have informative
purpose only.