Add shared payload handler method to resolve type
Type of the handler is enum shared with the default_handler
configuration. There is no need to have two different enums for this on
the other hand I want to keep convention of what is where in the module.
For that reason the new enum is inheriting the configuration enum.

Handler type will be propagated on the shared interface for all the
handlers.