Remove support for boot arguments without 'inst.' prefix from Dracut
We have the recommendation to use 'inst.' prefix everywhere a few years now so
let's make the step further and support only the 'inst.' variants.

This way we will finally solve conflicts with other kernel boot argument
consumers.

This is an accepted change by FESCO
https://fedoraproject.org/wiki/Changes/Ignore_Anaconda_kernel_boot_parameters_without_inst_prefix

Warn users that they should use the 'inst.' prefix variants and that the old
versions were removed.