Run cockpit-client instead of the Anaconda GUI
This is achieved by creating a new UI type called
cockpit, for the time being living in parallel with
the existing GUI and TUI implementations.

But unlike the existing UIs, the cockpit UI will just
run the cockpit-desktop web view executable.

Also as we are adding a new module, we need to adjust our
autotools makefiles accordingly.