Make the run() method of the Web UI blocking
Both our existing UIs run their main loop in the run() method,
resulting in blocking as long as the UI main loop is running.

In our case we just start the web-view application in the run method,
so also add a very long sleep() call to emulate the expected run()
blocking behavior.

This seems to fix both the need to switch to tty6 manually as well
removes the need to pass inst.nokill on boot command line.