Make possible to use fork correctly for running tests
During the container build we are using link to a repository to download spec
file to get list of the dependencies to be installed. Unfortunately, we can't
just use the one from the repository because container build cannot reach
anything outside of the build directory.

Because of that let's introduce a new ARG to our containers specifying where the
repository is. This will work only for GitHub but I don't think that should be a problem.