infra: Do not preserve metadata for Live builds
When copying RPM from the Anaconda RPM scratch builds we used to
preserve metadata. The issue is that there could be a wrong SELinux
context and uid/gui on these files.

That could result in broken Live ISO builds by Lorax in a weird way.
Without preserving the metadata we have a clean local packages which can
be easily consumed. Also there is no need to do this (to my knowledge).