Add side_payload to PayloadBase object
Side payload will allow us to link dependent payloads and resolve our
long standing issue of missing support for multiple payloads.

This way main payload can control it's side_payload.

Also side payload can have another side payload but that is not
supported currently.

----

This change will generalize handling of the flatpak_payload in the
payloads module.
This will allow us to resolve correct reaction of the Flatpak on
the DNF payload change after the package selection change. Which will
allow us to move space calculation to Task. That solves issue with
potential block of the UI because Flatpak payload is downloading
metadata.

Resolves: RHEL-75804