Change `NoSourceError` to `SourceSetupError`
The Flatpak source handling introduced local exception for
`NoSourceError` which should have been processed and used only locally.
However, code which can raise this exception is used in the installation
Tasks which could raise this exception outside the DBus module. Let's
instead use one of the standard errors we already have for payload.